Doc: unify "Dynamically Generating Menu Items" across modules
Both Qt Quick Controls and Qt Labs Platform can use the exact same
snippet for this, so use the same approach in both modules.

I would try to reuse the code (and documentation), but I'm not sure if
the linking to the types contained within would work, nor if there
would be conflicts, so to err on the side of caution, give each module
its own copies.

Task-number: QTBUG-64554
Change-Id: Id1cac733ad3eda12074234f882eeae289caa438a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>