sg: Expose the projection matrix on the QSGRenderNode itself
Turns out that implementing "proper" QSGRenderNode subclasses is
not quite straightforward in Qt 6. If a rendernode wants to
query the projection and modelview matrices, and store something
calculated based on them somewhere (e.g. push to a uniform buffer),
this is not actually possible to do in prepare(), only in render(),
and that is is no good when working with QRhi or some modern API.

For rendernodes that do not query the scenegraph's projectionMatrix
this is not a problem but for those that want it, there needs to be a
way to access it outside of the RenderState interface, because a
RenderState is only passed to render(), not to prepare().

[ChangeLog][QtQuick] Added a projectionMatrix() accessor to
QSGRenderNode.  This way a reimplemented prepare() can query both
matrices and can load the final modelview-projection matrix into a
uniform buffer for example. Previously this was only possible in
render() because the projection matrix was only exposed in the
RenderState. With QRhi, or modern APIs such as Vulkan, that is not
quite sufficient since it is more than likely that one will want to
calculate and store the matrix in prepare(), not in render().

Change-Id: Ic6ff734962381ee7b199329bf2a37b218c09dc2c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>