Tweak error message about non bindable properties
It is not necessarily the NOTIFY that is missing, e.g. I recently
saw an example that was missing the READ property declaration
but the error message still said "you don't have a NOTIFY" and well, the
code did indeed have NOTIFY so it was a bit disconcerting until i
realized the READ was missing.

Change-Id: Ic5bd4c2be0cd6229ab8cc543be235bbd147ae818
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>