CMake: Place *_qmllint* targets into a separate FOLDER
Set the FOLDER property to *_qmllint* targets to the value of the newly
introduced global property QT_QMLLINTER_TARGETS_FOLDER.  The default
folder name is "QmlLinter".

This only has an effect if USE_FOLDERS is ON and the CMake generator
supports folders.

The all_qmllint is not moved into the QmlLinter folder, because users
might want to trigger it manually.

[ChangeLog][CMake] Introduced the QT_QMLLINTER_TARGETS_FOLDER global
property to specify the name of the FOLDER for qmllint-related targets.

Task-number: QTBUG-99808
Change-Id: I2b89afdeab386507f7c2db17aa011a3b0535f2b4
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>