TextEditor example: use the ColorDialog from qt quick dialogs
In 6.4 we're introducing the ColorDialog in the qt quick dialogs module.
This was the final dialog, that was missing from the module in previous
versions of qt.

This change will make use of the new dialog, instead of the qt labs
platform version, which is using the widgets color dialog as a fallback.

The eventual goal for this example should be to eliminate the need for
the qt labs platform module entirely.

Note that I am also making a small change to how the font dialog is
being used, since we deprecated the currentFont property in the 6.3
release, but I forgot to update this example along with it.

Change-Id: Ib99ca7cbf110d3084ff96dfadfa1cd8c8af8815e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 314cae87524993a2af8068aa81c90ab2f1006590)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>