Allow implicit mouse grab by accepting only on press
If we install a DragHandler on a Controls Button and try to drag it:
on press, the DragHandler takes a passive grab, and the Button accepts
the event, which gives it an implicit exclusive grab. Then when we drag
past the drag threshold, the DragHandler takes over the exclusive grab;
but we also deliver the mouse move to the Button, because it already had
the exclusive grab before. It doesn't ignore() the event, but we do not
want to interpret the accepted state as a reason for the Button to
re-grab the mouse, because it's just a move, not a press. This should
restore the behavior to what we had in Qt 5.

Task-number: QTBUG-105610
Change-Id: I77e185b2737da19848e206cf08fc62cf62930d92
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 68dc224920ddf3aeb67a87d9b983107a88fd7974)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>