CMake: Add iOS build auto test for Qt Quick project
Currently we can't run iOS tests in the CI, but we can at least build
them.

Reuse the CMake deployment auto test as a regular build test
(no deployment) when targeting iOS.

This ensures we don't regress in building iOS projects that use qml
modules.

Move the deployment test to be added earlier in the CMakeLists.txt
file and return early when both IOS and QT_BUILD_MINIMAL_STATIC_TESTS
are true, so that we build only that test in the CI for iOS.

Pick-to: 6.4
Fixes: QTBUG-96056
Change-Id: I68002c5b70bb3f182cf55ec4a2d4888ea9e2b81e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>