iOS style: increase the tap area around the TreeViewDelegate indicator
As it stood, it could be quite hard to tap on an indicator to
expand or collapse the tree. So increase the size of the area
around it. Also make it independent of the arrow image size, since
the image will change size depending on if the indicator is
expanded or not.

Pick-to: 6.4
Change-Id: Idf7f9b6429fd5149a8fc262edc73a127740388c5
Reviewed-by: Doris Verria <doris.verria@qt.io>