Material: add TreeViewDelegate.qml
Add a version of TreeViewDelegate.qml to the Material
style, that looks more at home for that style (rather
than using the one in the Basic style).

This version is using the arrow-indicator from the
Material style as expand/collapse indicator, and
makes the height of each row match the height of a
normal Material button.

Change-Id: I084532ca6e527593e5711d66aa53f240f0356d99
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 55fab9a73e720a84d0f0a86cc4175f34a4573188)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>