TapHandler: don't cancel on release
The previous fix for QTBUG-94012 was in conflict with the idea that
TapHandler could be used to augment behavior on other event-handling
Items: it would emit canceled rather than tapped whenever there was a
rival exclusive grabber, unless that grabber was an event-filtering
item. The bug I was trying to fix was in the Qt Quick 3D dynamictexture
example: if you click on a TextEdit in a subscene to set the cursor
position, a TapHandler on the entire View3D should not react, because
something in the subscene was already reacting. The event accepted flag
also does not stop propagation in other subscenes. This bug should
rather be fixed by changing gesturePolicy so that the outer TapHandler
does try to take the exclusive grab, and then it is stolen by TextEdit
because TapHandler's default grabPermissions include
ApprovesTakeOverByAnything.

Handlers in general are useful to augment existing behavior of Items.
At least gesturePolicy and grabPermisions are available to fine-tune the
stealing of the exclusive grab; but when TapHandler is using a stealthy
passive grab (which it does with the default gesturePolicy), we don't
want to prevent it from reacting.

This reverts the TapHandler changes from commit
d342b8f77658537c4fd3318982d2ae7a964b6426 but keeps the autotest, which
is now improved by checking multiple gesturePolicy values and checking
for the canceled signal if the policy specifies an exclusive grab.

Also call QEventPoint::setAccepted(false) if the gesturePolicy is
DragThreshold, because if the TapHandler is getting by with a passive
grab, it should not interfere with delivery to other items. This makes
touch behavior more consistent with mouse behavior.

Pick-to: 6.4
Fixes: QTBUG-99887
Fixes: QTBUG-105609
Task-number: QTBUG-94012
Change-Id: Ib8f80d4e6523b01ef204a69f1f30f010bee6eb8f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>