Fusion: don't set ButtonText palette role
After 9c66af1f1d2cb8d25546d544e22fd7647227214e, we get correct palette
colors in both light and dark mode from qt_fusionPalette(), and so
the changes that QQuickFusionTheme does to the palette no longer appear
to be necessary. The values it set for ButtonText roles caused text to
be illegible.

Fixes: QTBUG-99276
Pick-to: 6.4
Change-Id: Ia35522c3676a8a544cd555690e0a1993a55aac70
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>