examples, texteditor: remove the config file, and use default settings instead
The qtquickcontrols2.conf file was hard coding the style to be
Material. In addition, it was overriding both the Primary and the
Foreground color, and sat Theme=System. The latter meant that the
application would run in either Light or Dark mode, depending on
the OS. And for Dark mode, the assigned colors caused the text to
end up with the same color as the background color, and therefore
be invisible.

The same with the toolbar buttons; since the icon color of a toolbar
button map to the primary color, and the primary color was hard-coded
to be the same as the background color in the configuration file, the
buttons would appear hidden.

Since we now have native styles for both macOS and Windows (and
Fusion on Linux), overriding the style (and especially the colors)
in a config file seems unnecessary.

This patch will therefore remove the whole config file, and use
the default style settings instead.

Fixes: QTBUG-105860
Change-Id: I330fe30746096cf26d8dca343f2c122d7c32c7fe
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit ec582ef0b9e31c5e8fb9d36a17f1e39f593d7455)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>