Clear delivery agent delegated focus
When focus was delegated, the focused item would remain as an active
focus item in the delivery agent after deletion. Further use by the
delivery agent would then cause a crash. But since
543598a6cc07d67e7651c9f65c058465ea6d8425 the intention is that the
window's DA's activeFocusItem is the one that counts (if there is only
one keyboard, there can be only one place to type text; and incoming
events will visit the window's DA first.) So any time we clear focus,
the window's DA has to clear it, just as setFocusInScope() delegates to
the window's DA.

Pick-to: 6.4 6.2
Fixes: QTBUG-105192
Change-Id: Iec8c6c67ff18b5dac5ec13fcced6e3fe30423c14
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>