Drop unnecessary creation contexts from Quick Controls
Bound components can only be instantiated in the context they're
declared in. Adding a context in between before instantiating a
component makes it impossible to bind the component. We want to use
bound components so that we can safely use IDs of other objects from
the same context inside the objects created from the component.

Pick-to: 6.4
Fixes: QTBUG-106042
Change-Id: I7a0e1ea3e079ccd0f5fe156f07f8bc62149c6c0a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>