Fix thread issue in TextEdit when font db updates
This amends 2fb43a35dfa08bb0746913e3363f1fcf21815e72, which
introduced invalidation of text when the font database changes.

The text invalidation is done from the render thread, which
mostly works well because it is done during sync, so the main
thread is locked. However, for TextEdit, the invalidation is
more complex and would implicitly call documentChanged() which
would then try to stop a timer. Since the timer has to be
stopped from the owner thread, it would fail and throw a
warning.

So for TextEdit, we invoke the method on the main thread
instead, which means it will use one extra update pass
in order to properly invalidate its contents.

Pick-to: 6.4
Fixes: QTBUG-105992
Change-Id: Idc020c682dcb2d1f3c66d043cf98fe9c6e4dc2a1
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>