iOS Style: Make the theme property read-only
We had marked the iOS Style's theme property as a writable property.
However, setting the theme through it was not supported. It also
doesn't make sense to set the theme through this attached property
since in the QPA layer we set the appearance(theme) and subscribe to
theme changes based on the traitCollection property of the UIWindow,
which changes based on system-wide configurations.

Change-Id: I36191a47d8892dff3069901e16abad82b9f9f82e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit c0f50622c7d75915e1a2f167b9bcfe6702f5dcb1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>