Imagine: fix ComboBox's implicit size bindings
The parentheses were misplaced, causing e.g. contentItem.implicitWidth
to be ignored. Without any parentheses, the same issue is present due
to evaluation order, so we do still need them, just in the correct
place.

While we're at it, we replace contentItem.implicitWidth with
implicitContentWidth so that Imagine respects
implicitContentWidthPolicy (which was added in
49ffc6e6af83b295c67fd119b79c925879cc292e).

Fixes: QTBUG-106098
Pick-to: 6.2 6.3 6.4
Change-Id: I6ff08f5cb9309ae0494d95917ce8b7a22d2ba663
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>