Binding: Restore state also when binding is deleted
When using a dynamically created Binding element (e.g. via a Loader or
createComponent), the Binding might get destroyed while its target is
still in continued use. So far, wo wouldn't have restored the original
values in that case. Fix this by checking in QQmlBind's dtor whether it
is necessary to restore tha state and by calling eval if necessary.

We however skip resetting when the whole engine is in the process of
shutting down (and also when the target object is already gone, as we
can't do anything in that case anyway).

Fixes: QTBUG-106103
Change-Id: I1a40778536877a19c674ccafa4dfb63ec0fa4405
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit e2fa7ab91310ea74c30e9458dfbe20d257578659)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>