Re-enable auto testing of remote image downloading on linux
It seems those tests were disabled for linux because of QTBUG-45655,
but that task was closed as resolved years ago. Also, recently another
remote data item was added to the autotest, without being disabled for
linux, and still this test has not caused CI problems. So there does
not seem to be any reason to keep skipping those tests on linux.

Task-number: QTBUG-45655
Change-Id: I928fc666dac768b5ffa22944191665ef62c053c2
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>