QQuickTableView: avoid printing out incorrect warning
A warning would be printed out from TableView if you called
positionViewAtCell with a mode that affected either horizontal
or vertical alignment, but not both. The reason is that the
overridden version would just forward the call to
positionViewAtRow() and positionViewAtColumn(), without
checking if the given mode allowed for it first.

This patch will instead check the mode first, before
deciding which functions to call. This will stop
the warning from being incorrectly printed.

Pick-to: 6.4
Change-Id: I58145873def7543fcffd51ade7d15c70fb755aae
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>