Do not listen to baselineOffset changes unless really needed
For some controls, (e.g. TextInput and Text) we can get changes in
baseline offset when they change their height. These changes will
invalidate the layout and generate a polish event.

This means that during a rearrange of a layout with TextInput children, it
might change the height of its children. This might change the baselineOffset
of these children, which again will cause the layout to have to be invalidated
(and rearranged again) This double-loop is unnecessary if the layout is not
supposed to align these items by their baselines, so we add some conditions to
see if we really need to connect to the baselineOffsetChanged() signal.

This also fixes a bug observed in the color dialog, where the
semi-recursiveness described above lead to a problem with that the alpha
TextField of the RGBA color input was *not* laid out by the parent RowLayout

Pick-to: 6.4 6.3 6.2
Fixes: QTBUG-105899
Change-Id: If66b4bbde6a37c962b157e166c883ba20ff24c21
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>