V4: Account for the guard pages when allocating stack space
Previously we've assumed the whole allocation can be used, even though
the first and the last page are actually not usable. This makes a
difference when the size of the guard pages grows, such as on macOS,
which these days has 16k pages.

Add the extra guard page size to the amount of memory to be allocated in
order to fix the calculation.

Fixes: QTBUG-93188
Change-Id: I0ebece94449da3127e9a78a19d8a22722ad8d698
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 826b77c8cf0ffbef4f95e7b9e72eb9dc25936657)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>