Flickable: don't initiate grab if there are multiple touchpoints
Flickable is meant to be flicked with one finger, not two or more.
If the UI depends on using the pinch gesture on some child of the
contentItem, Flickable shouldn't interfere with that.

The user might accidentally touch the screen with extra fingers while
trying to flick; Flickable should not care, as long as it has grabbed
the one touchpoint that it's tracking before the second point appears.

Removes QEXPECT_FAILs added in d7623d79ef4bc9533fced027bf1d173d68b4eba6

handleMoveEvent() can be called from filterPointerEvent(), or from
mouseMoveEvent() in case of direct delivery; and filterPointerEvent()
will now avoid stealing the grab during multi-touch press or release as
well. The "ignoring multi-touch" categorized log message is emitted
twice while filtering a move event, but only once in the other cases.

Fixes: QTBUG-77629
Task-number: QTBUG-85278
Task-number: QTBUG-106223
Fixes: QTBUG-109655
Pick-to: 6.2 6.4 6.5
Change-Id: I7a19deaed3b2255de2b686f0d96bd08fe5ab1636
Reviewed-by: Doris Verria <doris.verria@qt.io>