QQmlIncubatorPrivate: Undo old MSVC workaround
and use the chance to pack the class a bit more compactly: We don't need
to have Incubator as a direct subclass of QSharedData; instead, only
make QQmlIncubatorPrivate a subclass of it, and move the subComponentToCreate
into the alignment gap that exists on 64 bit systems. Moreover, reduce
the size of the Progress enum, and put it next to the boolean member.

Change-Id: I9ecfd9a779070c40bd8c375444e1faf307ba6fc4
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>