Fix visibility of scroll bars in ScrollView
When imperatively assigning a contentItem, we need to make sure that
we respect the policy of the scroll bars, otherwise they could end up
visible when they shouldn't be, and vice versa.

This patch also sets the ScrollView as the parent item of the
contentItem, as a ListView assigned imperatively may already have a
parent item, and so setContentItem_helper will not set one.
When setting the parent item, we also need to ensure that the
stacking order is correct, otherwise the scroll bars will be
behind the contentItem and can't be interacted with.

Task-number: QTBUG-106118
Pick-to: 6.2 6.5
Change-Id: I88d342282ea6d202cbe0f776a6bf4fb1f9cadc48
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>