Doc: Remove usages of "\js"
QDoc provides a "\js"/"\endjs", undocumented, unsupported, command pair
to allow writers to write javascript code-blocks in the documentation.

This command is part of a never-finalized javascript support for QDoc
that is now being removed.

To avoid breakages in the documentation, the usages of the command are
now removed.

The commands were replaced by a "\qml"/"\endqml" pair, that should be
mostly equivalent, albeit some highlighting may be lost in the offline
documentation.

Task-number: QTBUG-106275
Change-Id: I765c72542ff46320332d1535fed823421514ed30
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>