doc: Clarify QQuickShape renderer documentation
There is no direct OpenGL in the scenegraph any more, and even if
there were, it wouldn't make sense to differentiate here as long
as the renderer is the same.

Also remove a note which makes it sound like there are vendor-specific
extensions supported for shapes. This is no longer the case in Qt 6.

As a drive-by, I also removed the phrase "manual triangulation", as
I think this makes it sound like the user has to manually triangulate
the path. If anything, it should be automatic triangulation? I don't
think it's needed anyway.

Change-Id: I2ad7e7479416f6e69e8087954a30c720bd02b3ed
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>