Fix alias to inline component type
When we create an alias property, we resolve its targets metatype. In
the case of inline components of the same file, that's however not
possible via QQmlType, as the type will only be available once the outer
type has been fully set up.
Fix this by looking up the type in the ExecutableCompilationUnit, where
the information is available while the type registraiton process is still
ongoing.

As a drive-by, also set the alias flag correctly when constructing the
QMetaObject of a QML element.

Pick-to: 6.2 6.3 6.4
Fixes: QTBUG-106392
Change-Id: Ie8f55dd0894cc5c8d683dd3c685980878956d3ca
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>