qqmlprivate.h: make static constexpr members c++11 compliant
Newest gcc versions (latest 11 and 12) decided to be c++11 compliant for
static constexpr members, meaning that they need a redundant definition
at namespace scope. This behavior is required for c++11 but deprecated
and not required starting with c++17.
Added those definitions for all static constexpr members in
qqmlprivate.h with appropriate comment.

Fixes: QTBUG-106377
Change-Id: Idb3b3d38218f3a66908de259ae765db2aeb1febb
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>