Doc: make QQuickWindow::closing signal private
This was apparently never meant to be public, and can only be used via
the old signal syntax with a slot that takes no arguments, as trying to
use the (private) QQuickCloseEvent type will result in a build error.

Since the close event can't be accessed publicly from C++, it can't be
used to reject the event, making this a misleading and frustrating
experience for users.

So, mark it as internal.

It is possible to prevent the application from quitting by using an
event filter on the window or using the QML API.

Task-number: QTBUG-55722
Change-Id: I854ed0baad972e061a6bccfcc7d339191caec254
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>