CMake: Use _qt_internal_get_tool_wrapper_script_path
Instead of relying on the variable QT_TOOL_COMMAND_WRAPPER_PATH to be
set, use the new _qt_internal_get_tool_wrapper_script_path function
to get the path to the tool script wrapper.

This allows an indirection layer to ensure the script wrapper is first
created, regardless of whether we are building Qt, a user project or
an example via ExternalProject in a prefix build.

Pick-to: 6.4
Task-number: QTBUG-90820
Task-number: QTBUG-96232
Change-Id: I87763fdcb3dbb337546b38ad8a1a6bd7f45cec21
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>